refactor(app-icon): add src and types folders
