--- packages/app-icon/AppIcon.d.ts
@@ -1,14 +1,0 @@
-export interface AppIconProps {
-    tag?: React.ReactElement | string;
-    color?: string;
-    size?: string;
-    branded?: boolean;
-    className?: string;
-    src?: string;
-    alt?: string;
-    children?: any;
-}
-
-declare const AppIcon: React.FunctionComponent<AppIconProps>;
-
-export default AppIcon;
