+++ packages/app-icon/package.json
@@ -4,6 +4,7 @@
   "author": "Evan Sharp <evan.sharp@availity.com>",
   "description": "Availity UI Kit application icon react component.",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity"
@@ -26,5 +27,8 @@
   },
   "peerDependencies": {
     "react": ">=16.3.0"
+  },
+  "dependencies": {
+    "prop-types": "^15.7.2"
   }
 }
