+++ packages/app-icon/tsconfig.json
@@ -1,9 +1,9 @@
 {
-    "extends": "../../tsconfig",
-    "compilerOptions": {
-      "baseUrl": "../../",
-      "paths": {
-        "@availity/app-icon": ["packages/app-icon"],
-      }
-    },
+  "extends": "../../tsconfig",
+  "compilerOptions": {
+    "baseUrl": "../../",
+    "paths": {
+      "@availity/app-icon": ["packages/app-icon"]
+    }
   }
+}
