+++ packages/app-icon/types/AppIcon.d.ts
@@ -1,0 +1,13 @@
+export interface AppIconProps {
+  tag?: React.ReactElement | string;
+  color?: string;
+  size?: string;
+  branded?: boolean;
+  className?: string;
+  src?: string;
+  alt?: string;
+}
+
+declare const AppIcon: React.FC<AppIconProps>;
+
+export default AppIcon;
